Move main content from SidePanel.tsx to MainPage.tsx.
