+++ src/features/sidePanel/SidePanel/MainPage/MainPage.module.css
@@ -1,0 +1,166 @@
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@500&family=Titan+One&display=swap');
+
+.container {
+    width: 100%;
+    height: fit-content;
+    min-height: 100vh;
+}
+
+.loader {
+    position: absolute;
+    z-index: 9999;
+    top: 0;
+    left: 0;
+    height: 100%;
+    width: 100%;
+    background-color: white;
+    background-image: url("https://media.tenor.com/dZr67lNIXf8AAAAi/fish-animal.gif");
+    background-size: 300px;
+    background-repeat: no-repeat;
+    background-position: center;
+}
+
+.upperPanel {
+    background-color: var(--third);
+    display: grid;
+    grid-template-rows: auto 1fr auto;
+
+    header {
+        font-family: "Titan One", sans-serif;
+        font-size: 1.05rem;
+        background-color: var(--secondary);
+        color: white;
+        -webkit-text-stroke:  1px black;
+        border-bottom: solid 1px var(--primary);
+        width: 100%;
+        height: 3rem;
+        padding: 0.5rem;
+        display: flex;
+        align-items: center;
+        justify-content: space-between;
+        gap: 1rem;
+    
+        img {
+            object-fit: contain;
+            height: 100%;
+        }
+    }
+}
+
+.textareaContainer {
+    width: 100%;
+    height: 100%;
+    min-height: 0%;
+    
+    textarea, div {
+        font-family: "Poppins";
+        padding: 1rem;
+        width: 100%;
+        height: 100%;
+        overflow-y: scroll;
+        background-color: var(--third);
+        font-size: 0.7rem;
+        overflow-wrap: break-word;
+        resize: none;
+        border: none;
+        white-space: pre;
+        white-space: pre-wrap;
+    }
+
+    div {
+        background-color: #202020;
+        color: #E9E9E9;
+        scroll-behavior: smooth;
+
+        & > span {
+            background-color: var(--secondary);
+            color: #202020;
+            font-weight: bold;
+        }
+    }
+
+    textarea:focus-visible {
+        outline: none;
+    }
+}
+
+.buttonContainer {
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+    gap: 1rem;
+    height: 3rem;
+    padding: 0.5rem;
+    background-color: var(--secondary);
+    border-top: solid 1px var(--primary);
+
+    button {
+        font-size: 1rem;
+        padding: 0.35rem 0.5rem;
+        border-radius: 0.5rem;
+        border: none;
+        background-color: #3C93D2;
+        color: white;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+    }
+
+    button:hover {
+        color: #3C93D2;
+        background-color: white;
+        cursor: pointer;
+    }
+
+    button:active {
+        color: black;
+    }
+
+    button:disabled, button[disabled] {
+        background-color: grey;
+    }
+}
+
+
+
+.resizeHandle {
+    height: 1rem;
+    background-color: var(--secondary);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border-top: solid 1px var(--primary);
+    border-bottom: solid 1px var(--primary);
+}
+
+
+.lowerPanel {
+    height: 100%;
+    width: 100%;
+    position: relative;
+}
+
+.overlay {
+    position:absolute;
+    top: 0;
+    left: 0;
+    z-index: 999;
+    width: 100%;
+    height: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    
+    p {
+        font-size: 2rem;
+        color: white;
+        padding: 1rem;
+        background-color: rgba(0, 0, 0, 0.558);
+        font-family: "Titan One", sans-serif;
+        -webkit-text-stroke-color: black;
+        -webkit-text-stroke-width: 1px;
+        border-radius: 1rem;
+        text-wrap: wrap;
+        word-break: break-all;
+    }
+}
